Move transaction status icons into FastTransactionStatusBadge
